refactor: move custom form fields into a package structure
