+++ manage_breast_screening/mammograms/forms/appointment_cannot_go_ahead_form.py
@@ -1,6 +1,6 @@
 from django import forms
 
-from manage_breast_screening.core.form_fields import (
+from manage_breast_screening.nhsuk_forms.fields import (
     CharField,
     ChoiceField,
     MultipleChoiceField,
