+++ manage_breast_screening/mammograms/forms/screening_appointment_form.py
@@ -1,6 +1,6 @@
 from django import forms
 
-from manage_breast_screening.core.form_fields import ChoiceField
+from manage_breast_screening.nhsuk_forms.fields import ChoiceField
 
 
 class ScreeningAppointmentForm(forms.Form):
