+++ manage_breast_screening/nhsuk_forms/fields/integer_field.py
@@ -1,0 +1,29 @@
+from django import forms
+
+
+class IntegerField(forms.IntegerField):
+    def __init__(
+        self,
+        *args,
+        hint=None,
+        label_classes=None,
+        classes=None,
+        **kwargs,
+    ):
+        kwargs["template_name"] = "forms/input.jinja"
+
+        self.hint = hint
+        self.classes = classes
+        self.label_classes = label_classes
+
+        super().__init__(*args, **kwargs)
+
+    def widget_attrs(self, widget):
+        attrs = super().widget_attrs(widget)
+
+        # Don't use min/max/step attributes.
+        attrs.pop("min", None)
+        attrs.pop("max", None)
+        attrs.pop("step", None)
+
+        return attrs
